+++ app/components/pool/action/add/app-license-picker/app-license-picker.scss
