+++ app/components/pool/action/add/app-license-picker/index.ts
@@ -1,0 +1,1 @@
+export * from "./app-license-picker.component";
