+++ app/components/pool/action/add/index.ts
@@ -1,8 +1,8 @@
-export * from "./app-license-picker.component";
+export * from "./app-license-picker";
 export * from "./container-configuration.component";
 export * from "./container-images-picker.component";
 export * from "./container-registry-picker.component";
-export * from "./license-eula-dialog.component";
+export * from "./license-eula-dialog";
 export * from "./pool-create-basic-dialog.component";
 export * from "./pool-os-picker.component";
 export * from "./os-offer-tile.component";
