+++ app/components/pool/action/add/license-eula-dialog/index.ts
@@ -1,0 +1,1 @@
+export * from "./license-eula-dialog.component";
